django_2d_array/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
pytest==5.4.3
tox==3.15.2

